refactor: move dork results utility and improve error handling in filters
- Relocated the get_and_save_dork_results function from misc.py to external.py for better modularity.
- Enhanced error logging and parsing logic in the parse_references template filter to provide more detailed feedback on parsing failures.
- Improved input validation in the ToggleSubdomainImportantStatus API view by checking for missing subdomain_id.
- Removed the unused FetchSubscanResults API view for codebase simplification.
- Cleaned up imports in scan.py by removing an unused import.
